// src/worker/slack/schema.ts
import * as z from "zod";

export const slashCommandSchema = z.object({
  token: z.string(),
  team_id: z.string(),
  team_domain: z.string(),
  channel_id: z.string(),
  channel_name: z.string(),
  user_id: z.string(),
  user_name: z.string(),
  command: z.string(),
  text: z.string(),
  response_url: z.string(),
  trigger_id: z.string(),
});

export type SlashCommand = z.infer<typeof slashCommandSchema>;